Fixed gpg-agent setup
Apparently it wasn't doing what I thought
it was. The gpg2 formula comes with a copy
of gpg-agent which is the same version as
itself. The standalone gpg-agent package
was actually quite outdated.

Also, the start script wasn't running correctly
and so the older gpg-agent was never being used.
So I didn't even know something was wrong.