Added module aliasing, install Node packages
An installer can be an alias for another. The installer
can detect when it's called from an alias by looking at
its ModuleName param. It can then change functionality
based on the alias name such as the node installer.

Like Python and system packages, the node installer
will globally add a set of packages as defined in
the settings file.