More bspwm and polybar setup
Polybar is finished (for now).
bspwm is mostly finished. I'm sure there will be some
finner tweaks as I start to use it. Same with the keybinds.

The scratchpad is really giving me fits. If I find time I need
to look at a more robust way to manage it.